rag/langextract-rag/pyproject.toml
[project]
name = "langextract-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langextract>=1.0.8",
    "python-docx>=0.8.11",
    "mammoth>=1.6.0",
    "six>=1.16.0",
    "python-dateutil>=2.8.0",
    "pandas>=2.0.0",
]
